pyproject.toml
[tool.poetry]
name = "ViSQL"
version = "0.1.0"
description = "ViSQL - Natural Language to SQL + Visualization AI System"
authors = ["Pritam <pritampaul010@gmail.com>"]
readme = "README.md"
packages = [{ include = "ViSQL" }]
# Optional
# homepage = "https://your.project.site"
# license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.119.0"
uvicorn = {extras = ["standard"], version = "^0.37.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
asyncpg = "^0.30.0"
aiosqlite = "^0.21.0"
pydantic = {extras = ["email"], version = "^2.12.3"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
bcrypt = "<4.0"
argon2-cffi = "^25.1.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "httpx[httpx] (>=0.28.1,<0.29.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
